+++ Tests/CleanArchitecture.Comprehensive/CleanArchitecture.Comprehensive.Infrastructure/Persistence/Configurations/Pagination/PersonEntryConfiguration.cs
@@ -1,0 +1,26 @@
+using CleanArchitecture.Comprehensive.Domain.Entities.Pagination;
+using Intent.RoslynWeaver.Attributes;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.EntityFrameworkCore.EntityTypeConfiguration", Version = "1.0")]
+
+namespace CleanArchitecture.Comprehensive.Infrastructure.Persistence.Configurations.Pagination
+{
+    public class PersonEntryConfiguration : IEntityTypeConfiguration<PersonEntry>
+    {
+        public void Configure(EntityTypeBuilder<PersonEntry> builder)
+        {
+            builder.HasKey(x => x.Id);
+
+            builder.Property(x => x.FirstName)
+                .IsRequired();
+
+            builder.Property(x => x.LastName)
+                .IsRequired();
+
+            builder.Ignore(e => e.DomainEvents);
+        }
+    }
+}