+++ Modules/Intent.Modules.Application.MediatR.CRUD/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 6.0.17
 
 - Fixed: Issue where `CRUD Update Action` was not generating correct code if Domain Services were injected.
+- Fixed: Query with nullable mapped property, which is paginated generates uncompilable code.
 
 ### Version 6.0.16